Add mass changes.
Includes massive changes. Checking out only this commit will probably
not result in working system.
This commit is a result of not committing changes for a long time and
being checked out in an unrelated branch. All new files in this commit
are the result of a stash merge.

Below are some of the summarized changes:

	- Include public assets
	- Update gitignore
	- Add new controller handling various domains
	- New DI container file
	- routes file and database configuration file
	- update frontcontroller
	- Update views

The handlers directory will be removed in the future.